+++ scripts/src/form/interfaces.d.ts
@@ -58,13 +58,6 @@ interface Data {
     [propName: string]: any;
 }
 
-interface IResponse {
-    value: any;
-    formId: string;
-    resId: string;
-    [propName: string]: any;
-}
-
 interface IFormPageState {
     schema: Schema,
     schemaMetadata: SchemaMetadata,